requirements.txt
python-telegram-bot
pillow
imageio
nest_asyncio
flask
gunicorn

